% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fivethirtyeightdata-package.R
\docType{package}
\name{fivethirtyeightdata}
\alias{fivethirtyeightdata}
\title{fivethirtyeightdata: Data and Code Behind the Stories and Interactives at 'FiveThirtyEight'}
\description{
An add-on R data package that contains the large datasets not found in the original
\code{fivethirtyeight} package via a drat repository \url{https://github.com/fivethirtyeightdata/drat}
published by FiveThirtyEight \url{https://github.com/fivethirtyeight/data}. Note
that while we received guidance from editors at 538, this package is not
officially published by 538. Contribute to this package at
\url{https://github.com/fivethirtyeightdata/fivethirtyeightdata}.
}
\examples{
# Example usage:
library(fivethirtyeightdata)
head(senators)

# All information about any data set can be found in the help file:
?senators

# To view a list of all data sets:
data(package = "fivethirtyeightdata")
}
