% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fivethirtyeightdata.test-package.R
\docType{package}
\name{fivethirtyeightdata.test}
\alias{fivethirtyeightdata.test}
\title{fivethirtyeightdata.test: Data and Code Behind the Stories and Interactives at 'FiveThirtyEight'}
\description{
An add-on R data package that contains the large datasets not found in the original
\code{fivethirtyeight} package via a drat repository \url{https://github.com/mariumtapal/drat}
published by FiveThirtyEight \url{https://github.com/fivethirtyeight/data}. Note
that while we received guidance from editors at 538, this package is not
officially published by 538. Contribute to this package at
\url{https://github.com/mariumtapal/fivethirtyeightdata.test}.
}
\examples{
# Example usage:
library(fivethirtyeightdata.test)
head(all_senators)

# All information about any data set can be found in the help file:
?all_senators

# To view a list of all data sets:
data(package = "fivethirtyeightdata.test")
}
